enable ffmpeg-full extension
Adding this extension which is built by the SDK (so; the ffmpeg ABI matches the less capable one bundled with the runtime) will provide a fully-featured ffmpeg version to your app. It will be installed automatically, but users who uninstall it or choose not to use it will be able to continue as-is with the default bundled ffmpeg.