clean up and delete modules available in the runtime
`gstreamer` is already available in the runtime, we just need to build
`gst-editing-services`.

As for the other encoders/decoders, they are either already available in
the runtime or should be separately packaged as a plugin for Footage. A
runtime plugin like `org.freedesktop.Platform.codecs-extra` and
`org.freedesktop.Platform.openh264` could be used also to provide them.